Add missing db procedures
