+++ .changeset/plenty-lobsters-boil.md
@@ -1,0 +1,8 @@
+---
+"@neo4j/cypher-builder": patch
+---
+
+Add support for missing fulltext procedures:
+
+-   `db.index.fulltext.awaitEventuallyConsistentIndexRefresh`
+-   `db.index.fulltext.listAvailableAnalyzers`
