+++ .changeset/red-guests-hunt.md
@@ -1,0 +1,11 @@
+---
+"@neo4j/cypher-builder": patch
+---
+
+Add support for missing db procedures:
+
+-   `db.ping`
+-   `db.propertyKeys`
+-   `db.relationshipTypes`
+-   `db.resampleIndex`
+-   `db.resampleOutdatedIndexes`
