+++ src/namespaces/db/db.ts
@@ -18,8 +18,7 @@
  */
 
 import { CypherFunction } from "../../expressions/functions/CypherFunctions";
-import type { VoidCypherProcedure } from "../../procedures/CypherProcedure";
-import { CypherProcedure } from "../../procedures/CypherProcedure";
+import { CypherProcedure, VoidCypherProcedure } from "../../procedures/CypherProcedure";
 import type { Expr } from "../../types";
 import { normalizeExpr } from "../../utils/normalize-variable";
 
@@ -29,13 +28,7 @@ export * as schema from "./schema";
 
 const DB_NAMESPACE = "db";
 
-/** Returns all labels in the database
- * @see [Neo4j Documentation](https://neo4j.com/docs/operations-manual/5/reference/procedures/#procedure_db_labels)
- * @group Procedures
- */
-export function labels(): CypherProcedure<"label"> {
-    return new CypherProcedure("labels", [], DB_NAMESPACE);
-}
+// FUNCTIONS
 
 /**
  * @see [Neo4j Documentation](https://neo4j.com/docs/cypher-manual/current/functions/database/#functions-database-nameFromElementId)
@@ -46,12 +39,25 @@ export function nameFromElementId(dbName: Expr | string): CypherFunction {
     return new CypherFunction("nameFromElementId", [dbNameExpr], DB_NAMESPACE);
 }
 
+// PROCEDURES
+
 /**
- * @see [Neo4j Documentation](https://neo4j.com/docs/operations-manual/current/reference/procedures/#procedure_db_info)
- * @group Procedures
+ * @see [Neo4j Documentation](https://neo4j.com/docs/operations-manual/current/reference/procedures/#procedure_db_awaitindex)
+ * @group Functions
  */
-export function info(): CypherProcedure<"id" | "creationDate"> {
-    return new CypherProcedure("info", [], DB_NAMESPACE);
+export function awaitIndex(indexName: Expr | string, timeOutSeconds: Expr | number): VoidCypherProcedure {
+    const indexNameExpr = normalizeExpr(indexName);
+    const timeOutSecondsExpr = normalizeExpr(timeOutSeconds);
+    return new VoidCypherProcedure("awaitIndex", [indexNameExpr, timeOutSecondsExpr], DB_NAMESPACE);
+}
+
+/**
+ * @see [Neo4j Documentation](https://neo4j.com/docs/operations-manual/current/reference/procedures/#procedure_db_awaitindexes)
+ * @group Functions
+ */
+export function awaitIndexes(timeOutSeconds: Expr | number): VoidCypherProcedure {
+    const timeOutSecondsExpr = normalizeExpr(timeOutSeconds);
+    return new VoidCypherProcedure("awaitIndex", [timeOutSecondsExpr], DB_NAMESPACE);
 }
 
 /**
@@ -60,7 +66,7 @@ export function info(): CypherProcedure<"id" | "creationDate"> {
  */
 export function createLabel(newLabel: Expr | string): VoidCypherProcedure {
     const newLabelExpr = normalizeExpr(newLabel);
-    return new CypherProcedure("createLabel", [newLabelExpr], DB_NAMESPACE);
+    return new VoidCypherProcedure("createLabel", [newLabelExpr], DB_NAMESPACE);
 }
 
 /**
@@ -69,7 +75,7 @@ export function createLabel(newLabel: Expr | string): VoidCypherProcedure {
  */
 export function createProperty(newProperty: Expr | string): VoidCypherProcedure {
     const newPropertyExpr = normalizeExpr(newProperty);
-    return new CypherProcedure("createProperty", [newPropertyExpr], DB_NAMESPACE);
+    return new VoidCypherProcedure("createProperty", [newPropertyExpr], DB_NAMESPACE);
 }
 
 /**
@@ -78,5 +84,60 @@ export function createProperty(newProperty: Expr | string): VoidCypherProcedure 
  */
 export function createRelationshipType(newRelationshipType: Expr | string): VoidCypherProcedure {
     const newRelationshipTypeExpr = normalizeExpr(newRelationshipType);
-    return new CypherProcedure("createRelationshipType", [newRelationshipTypeExpr], DB_NAMESPACE);
+    return new VoidCypherProcedure("createRelationshipType", [newRelationshipTypeExpr], DB_NAMESPACE);
+}
+
+/**
+ * @see [Neo4j Documentation](https://neo4j.com/docs/operations-manual/current/reference/procedures/#procedure_db_info)
+ * @group Procedures
+ */
+export function info(): CypherProcedure<"id" | "creationDate"> {
+    return new CypherProcedure("info", [], DB_NAMESPACE);
+}
+
+/** Returns all labels in the database
+ * @see [Neo4j Documentation](https://neo4j.com/docs/operations-manual/5/reference/procedures/#procedure_db_labels)
+ * @group Procedures
+ */
+export function labels(): CypherProcedure<"label"> {
+    return new CypherProcedure("labels", [], DB_NAMESPACE);
+}
+
+/**
+ * @see [Neo4j Documentation](https://neo4j.com/docs/operations-manual/current/reference/procedures/#procedure_db_ping)
+ * @group Procedures
+ */
+export function ping(): CypherProcedure<"success"> {
+    return new CypherProcedure("ping", [], DB_NAMESPACE);
+}
+
+/**
+ * @see [Neo4j Documentation](https://neo4j.com/docs/operations-manual/current/reference/procedures/#procedure_db_propertykeys)
+ * @group Procedures
+ */
+export function propertyKeys(): CypherProcedure<"propertyKey"> {
+    return new CypherProcedure("propertyKeys", [], DB_NAMESPACE);
+}
+
+/**
+ * @see [Neo4j Documentation](https://neo4j.com/docs/operations-manual/current/reference/procedures/#procedure_db_relationshiptypes)
+ * @group Procedures
+ */
+export function relationshipTypes(): CypherProcedure<"relationshipType"> {
+    return new CypherProcedure("relationshipTypes", [], DB_NAMESPACE);
+}
+
+/**
+ * @see [Neo4j Documentation](https://neo4j.com/docs/operations-manual/current/reference/procedures/#procedure_db_resampleindex)
+ * @group Procedures
+ */
+export function resampleIndex(): CypherProcedure<"indexName"> {
+    return new CypherProcedure("resampleIndex", [], DB_NAMESPACE);
+}
+/**
+ * @see [Neo4j Documentation](https://neo4j.com/docs/operations-manual/current/reference/procedures/#procedure_db_resampleoutdatedindexes)
+ * @group Procedures
+ */
+export function resampleOutdatedIndexes(): VoidCypherProcedure {
+    return new CypherProcedure("resampleOutdatedIndexes", [], DB_NAMESPACE);
 }
